App commited
An one-page that uses the IMDB api to fetch any movie you search for, and displays all of the records of that movie title with the year it was produced. 
Very simple React functionality making use of hooks better, useState  and useEffects, also using props better to transfer data dynamically through components.